split up create code into separate files
