+++ src/matrix/room/common.ts
@@ -25,3 +25,18 @@ export const REDACTION_TYPE = "m.room.redaction";
 export function isRedacted(event) {
     return !!event?.unsigned?.redacted_because;
 }
+
+export enum RoomStatus {
+    None = 1 << 0,
+    BeingCreated = 1 << 1,
+    Invited = 1 << 2,
+    Joined = 1 << 3,
+    Replaced = 1 << 4,
+    Archived = 1 << 5,
+}
+
+export enum RoomType {
+    DirectMessage,
+    Private,
+    Public
+}
